bug: Fix undefined value when reading `response.data` in error block
The outer catch block recieves two types of error from unfulfilled promise. One is from the failed request to cloudinary when uploading a photo, the other is from the app server should the recipe object creation fails. Error object from cloudinary do not return an error with response key which differs from the error object sent from the web app server. Hence the need arises for the errprs to be handled differently.